pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "compressx"
version = "0.1.0"
description = "A modular LLM compression pipeline for Hugging Face models."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "accelerate==0.27.0",
  "click==8.1.7",
  "datasets==2.17.0",
  "fastapi==0.109.0",
  "huggingface-hub==0.20.3",
  "numpy==1.26.4",
  "pydantic==2.6.1",
  "psutil==5.9.8",
  "pyyaml==6.0.1",
  "requests==2.31.0",
  "safetensors==0.4.2",
  "torch==2.1.0",
  "tqdm==4.66.0",
  "transformers==4.38.0",
  "uvicorn==0.27.0",
]

[project.optional-dependencies]
distillation = [
  "peft==0.8.0",
]
quantization = [
  "auto-gptq==0.7.1",
  "bitsandbytes==0.42.0",
]
dev = [
  "httpx==0.27.0",
  "pytest==8.3.5",
]

[project.scripts]
compressx = "compressx.cli.compress:main"

[tool.setuptools.packages.find]
include = ["compressx*"]

[tool.pytest.ini_options]
testpaths = ["compressx/tests"]
asyncio_default_fixture_loop_scope = "function"
